dev/bonezone-model-support: cleanup
